requirements.txt
django==3.2.10
djangorestframework==3.13.0
django-phonenumber-field==6.0.0
django-allauth==0.13.0
django-rest-auth==0.9.5
phonenumbers==8.12.42
psycopg2-binary==2.9.3
psycopg2
pillow==8.4.0
djangorestframework-simplejwt==5.0.0
razorpay ==1.2.0
stripe==2.65.0
autopep8==1.6.0
autoflake==1.4
